LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/marker-animate-expected.txt


FAIL 'color' animation assert_equals: color expected "rgb(100, 50, 150)" but got "rgb(0, 0, 0)"
PASS 'opacity' animation 
FAIL 'color' + 'opacity' animation assert_equals: color expected "rgb(100, 50, 150)" but got "rgb(0, 0, 0)"
list item
